+++ max_div/solver/_strategies/__init__.py
@@ -1,1 +1,3 @@
-from ._base import SolverStrategy, StrategyType, hours, iterations, minutes, seconds
+from ._init_random import InitRandom
+from ._optim_dummy import OptimDummy
+from .core import SolverStrategy, StrategyDuration, StrategyType, hours, iterations, minutes, seconds